// apps/felexe-acquirer-service/src/auth/enums/audit-event.enum.ts
export enum AuditEvent {
  USER_REGISTERED = 'USER_REGISTERED',
  LOGIN_SUCCESS = 'LOGIN_SUCCESS',
  LOGIN_FAILED = 'LOGIN_FAILED',
  OTP_SENT = 'OTP_SENT',
  OTP_VERIFIED = 'OTP_VERIFIED',
  OTP_FAILED = 'OTP_FAILED',
  OTP_EXPIRED_DELETED = 'OTP_EXPIRED_DELETED',
  PASSWORD_UPDATE_REQUIRED = 'PASSWORD_UPDATE_REQUIRED',
  PASSWORD_UPDATED = 'PASSWORD_UPDATED',
  FORGOT_PASSWORD_REQUESTED = 'FORGOT_PASSWORD_REQUESTED',
  FORGOT_PASSWORD_COMPLETED = 'FORGOT_PASSWORD_COMPLETED',
}